mcp-blcwrtr/servers/preflight/requirements.txt
mcp-server-python==0.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
jsonschema==4.20.0
jinja2==3.1.2
pyyaml==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
